pyproject.toml
[project]
name = "valuebrief"
version = "0.1.0"
description = "Covering your assets. An automated daily digest tracking intrinsic value, margin of safety, and portfolio fundamentals."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.13.0",
    "certifi>=2026.2.25",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.6",
    "curl-cffi>=0.14.0",
    "datetime>=6.0",
    "ddgs>=9.12.0",
    "distro>=1.9.0",
    "dotenv>=0.9.9",
    "frozendict>=2.4.7",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna>=3.11",
    "jiter>=0.13.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.1.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-openrouter>=0.2.0",
    "openrouter>=0.7.11,<0.8",
    "langgraph>=1.1.3",
    "langgraph-checkpoint>=4.0.1",
    "langgraph-prebuilt>=1.0.8",
    "langgraph-sdk>=0.3.12",
    "langsmith>=0.7.22",
    "multitasking>=0.0.12",
    "numpy>=2.4.3",
    "openai>=2.30.0",
    "orjson>=3.11.7",
    "ormsgpack>=1.12.2",
    "packaging>=26.0",
    "pandas>=3.0.1",
    "peewee>=4.0.3",
    "pip>=26.0.1",
    "platformdirs>=4.9.4",
    "protobuf>=7.34.1",
    "pycparser>=3.0",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2026.1.post1",
    "pyyaml>=6.0.3",
    "regex>=2026.2.28",
    "requests-toolbelt>=1.0.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "soupsieve>=2.8.3",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "typing>=3.10.0.0",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "urllib3>=2.6.3",
    "uuid-utils>=0.14.1",
    "websockets>=16.0",
    "xxhash>=3.6.0",
    "yfinance>=0.2.66",
    "zope-interface>=8.2",
    "zstandard>=0.25.0",
    "markdownify>=1.2.2",
    "psycopg[async,binary,pool]>=3.3.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langchain-google-genai>=4.2.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
